+++ kuma/wiki/tests/conftest.py
@@ -10,6 +10,7 @@ import pytest
 from ..models import Document, DocumentZone, Revision
 
 
+BannedUser = namedtuple('BannedUser', 'user ban')
 Contributors = namedtuple('Contributors', 'valid banned inactive')
 DocWithContributors = namedtuple('DocWithContributors', 'doc contributors')
 DocHierarchy = namedtuple('DocHierarchy', 'top middle_top middle_bottom bottom')
@@ -17,6 +18,52 @@ KumaScriptToolbox = namedtuple(
     'KumaScriptToolbox',
     'errors errors_as_headers macros_response'
 )
+
+
+@pytest.fixture
+def wiki_user_2(db, django_user_model):
+    """A second test user."""
+    return django_user_model.objects.create(
+        username='wiki_user_2',
+        email='wiki_user_2@example.com',
+        date_joined=datetime(2017, 4, 17, 10, 30))
+
+
+@pytest.fixture
+def wiki_user_3(db, django_user_model):
+    """A third test user."""
+    return django_user_model.objects.create(
+        username='wiki_user_3',
+        email='wiki_user_3@example.com',
+        date_joined=datetime(2017, 4, 23, 11, 45))
+
+
+@pytest.fixture
+def inactive_wiki_user(db, django_user_model):
+    """An inactive test user."""
+    return django_user_model.objects.create(
+        is_active=False,
+        username='wiki_user_slacker',
+        email='wiki_user_slacker@example.com',
+        date_joined=datetime(2017, 4, 19, 10, 58))
+
+
+@pytest.fixture
+def banned_wiki_user(db, django_user_model, wiki_user):
+    """A banned test user."""
+    user = django_user_model.objects.create(
+        username='bad_wiki_user',
+        email='bad_wiki_user@example.com',
+        date_joined=datetime(2017, 4, 18, 9, 15)
+    )
+    ban = user.bans.create(by=wiki_user, reason='because')
+    return BannedUser(user=user, ban=ban)
+
+
+@pytest.fixture
+def create_revision(root_doc):
+    """A revision that created an English document."""
+    return root_doc.revisions.first()
 
 
 @pytest.fixture
